Lower AAC bitrate for voice message from 128 to 48 kbps
Good enough quality and less space/network used.

Initial intent is be consistent with Android (~36kbps), but as this seems to be stereo, give it a little more, so 48kbps. This quick change would lower storage usage by a factor > 2.6. Setting it to mono with even lower bitrate can also be an option.