Donations should only allow the server's currencies
If a currency is not in the list returned by the server, you shouldn't
be able to choose it because it won't work.

This worked correctly for monthly and gift donations, but not for
one-time. This fixes that.

This still has a bug where if your default currency isn't in the list,
you'll still be able to choose it. (Your default currency is locale-
-dependent; see `Stripe.defaultCurrencyCode`.) However, I plan to
address that in an upcoming change, and this commit doesn't make that
bug *worse*.