Add `FiatMoney` type
This change should have no user impact.

We commonly pair a currency, like USD, and an amount, like 1.23. This
adds the `FiatMoney` struct. It's a simple struct with two fields.

After adding it, I tried to use it everywhere. (It's possible I missed a
spot.)

I think this is a useful change on its own, but it'll be nice for an
upcoming change, too.

See also: [Android's equivalent class][0].

[0]: https://github.com/signalapp/Signal-Android/blob/cb65347bb387317856c1bbfa4b1d81080fd1024b/core-util/src/main/java/org/signal/core/util/money/FiatMoney.java#L1