Persist updates as part of fetch profile promise
As far back as 823927685d2c7da266054cd356cf755263883e18, the profile
updates were asynchronous. At that point it time, it didn’t matter
because there wasn’t a completion block; callers couldn’t know when the
updates *should* be complete, so it didn’t matter when they finished.

However, in 856fef76647c199122d160c5f52faddf230e2a0f a Promise result
was added, presumably to allow callers to know when it was safe to check
GV2 capabilities. Now, a cursory reading of the API would suggest that
all changes were persisted once the Promise was resolved.

This can result in bugs, such as [this example][0], where we fetch the
profile and then check a capability in the database. Even though it’ll
*probably* be the case, there’s no guarantee that the updated profile
will be persisted when we try to fetch the latest capability.

[0]: https://github.com/signalapp/Signal-iOS/blob/465d00664b46a73123f7e7d7c42e3f50d8608183/Signal/src/ViewControllers/AppSettings/Donations/BadgeGiftingConfirmationViewController.swift#L69-L88