Use RingRTC to validate 1:1 call offer messages in the NSE
This avoids an issue where the NSE might decide to wake up the main
app to respond to an Offer, but then the RingRTC in the main app
decides the ring has already expired. We've been careful about
exposing new RingRTC APIs to the NSE because we don't want to dirty
too much memory, but this particular one should be safe.