Check if payments are enabled after selecting user
There’s a TODO in the code about waiting for the profile fetch to
finish. As it turns out, the method we call after this one *also* checks
if payments are enabled, and if they’re not, it’ll do a profile fetch
before showing an error.

It will also provide a better error message if you select someone with a
missing profile key (and will show a “Send Message” button that was
previously unreachable).