Crash, don't throw, if donation receipt generation fails
_I recommend reviewing this with whitespace changes disabled._

We need to generate a "receipt request" before doing a donation. This is
is entirely on-device and is unlikely to fail, but it technically could.

Previously, `generateReceiptRequest` would throw if this failed. We did
a bad job handling these failures (in some cases, we just ignored the
failures completely!).

Because this represents such a catastrophic failure, we'd _like_ this to
crash the app so we can investigate it more urgently.