Introduce SnapshotManager
SnapshotManager's job will be to own two copies of
State: the "source of truth" and the
mainthread-accessible snapshot. It will also be
responsible for notifying a main-thread delegate
of updates. At the moment it has only one instance
because the snapshotting code hasn't been written
yet. However, it is already useful: it ensures
that all mutations funnel through a single
function while allowing read-only access to state
with minimal ceremony.