Donations: fix bug when double-tapping custom amount
Previously, the following bug was possible:

1. Go to the one-time donation view
2. Enter any custom amount, such as $5
3. Tap anywhere in the text field

This caused the Continue button to be disabled, which was wrong.

This happened because I wrongly assumed that "tap on text field" meant
"set the one-time amount to 0". Now, this is done more judiciously,
which fixes the bug.

(cherry picked from commit 30ef00342c702a7632bf0a0cc0748da09a300621)