Stop clobbering an argument when uploading debug logs
This change should have no user impact.

Here's a pseudocode version of the diff:

     def upload(data, mimeType):
    -  mimeType = "application/zip"
       doUpload(data, mimeType)

     upload([1, 2, 3], "application/zip")

This method needs to set a `Content-Type` for uploaded debug logs. It
accepts a `mimeType` argument but ignores it, instead hard-coding the
value to `application/zip`. This removes the hard-coded value and
respects the argument.

In practice, this argument was always `application/zip` anyway, so
this change has no effect. I think this is a useful fix on its own but
may be useful for an upcoming change.