Donation error string metadata should mention Apple Pay
This change should have no user impact.

`DONATION_PAYMENT_ERROR_*` is now `APPLE_PAY_DONATION_ERROR_*`, and
Apple Pay is now mentioned in the string comments. None of the
user-visible strings actually change.

We plan to show different errors for cards and PayPal and this will help
us get there, but I think this is a useful change on its own.