Update SQLCipher to 4.5.2
This updates SQLCipher from 4.5.1 to 4.5.2.

The changes listed in [SQLCipher's release notes][0] don't seem to
affect us except for the big change, which updates the "upstream" SQLite
from 3.37.2 to 3.39.2. [SQLite's changelog][1] lists a number of new
features and bug fixes.

I'm updating this to keep us on the latest version and pick up bug
fixes, not because there's a particular thing that needs fixing or a
feature we need.

[0]: https://github.com/sqlcipher/sqlcipher/blob/c1f02fd9689da4c27b4ce98be3d4cfb34db7980e/CHANGELOG.md#452---august-2022---452-changes
[1]: https://sqlite.org/changes.html