Fix nav bars during registration & change number
The new OWSNavigationController design handles showing/hiding the nav
bar. The code in this class was fighting against it, which caused the
nav bars to appear/disappear unexpectedly throughout the process.

The vc.navigationController.navigationItem.backBarButtonItem.isEnabled
code also appears to be a no-op. A view controller should configure its
own navigationItem (ie vc.navigationItem), not that of its
navigationController. Furthermore, backBarButtonItem isn't set to what's
actually shown -- this is an opportunity to override what should be
shown to our own custom button.