Fix onboardingMode assignments & checks
The OnboardingModeSwitchConfirmationViewController bases all of its UI
and behavior on the default mode for the current device. As a result,
the “toggle” behavior is wrong -- it should always override the mode.

Furthermore, if you back out of the override view controller and tap the
normal “Continue” button, that should use the default mode (because you
haven’t gone through the flow to confirm the overridden mode).