Don’t owsDebugFail if data items overlap mentions
If you mention somebody and we only know their phone number, the text of
the message will be something like “@+1 (650) 555-0100, how are you?” In
this case, we’ll have a mention from the beginning of the string to the
comma, and we’ll have a phone number match from the first character to
the comma. These overlap and trigger the error, but that’s fine.