Remove rarely-seen YDB onboarding screen
We used to use [YapDatabase] but [removed it in early 2021][1] after a
[migration to GRDB that started in 2019][2].

If you haven't opened the app since before we dropped YDB--in other
words, since early 2021--you'll see the following text on a screen
before you open the app:

> Because you've been inactive for a long period of time, you must
> register again to use Signal.

Tapping "Next" will send you into the normal registration flow.

I think this is a tiny number of users, so I removed this screen and
associated code. This will help simplify future onboarding code.

[1]: https://github.com/signalapp/Signal-iOS/commit/86b8eb08b8a67a4588bb3f397a2384ab285b08f3
[2]: https://github.com/signalapp/Signal-iOS/commit/2a5683b843346de517eaec3381e66ab1716ede8d
[YapDatabase]: https://github.com/signalapp/YapDatabase