Ensure that the video mute button in calls has the right initial state
Previously, if you answered a call on the lockscreen but hadn't given
Signal permission to use the camera, opening the app would show video
as unmuted but with no actual video being sent. By explicitly enabling
or disabling video when the call starts (and allowing the camera to be
considered "enabled" when the app is in the background, even though no
video will be sent until it returns to the foreground), the button is
always in the right state.