Double tap message to show context menu
In the conversation view, long pressing a message presents the context
view for that cell, allowing the user to add reactions, reply, delete
the message, etc. To match the behavior that iOS users are accustomed to
in iMessage, Signal should also allow double-tapping messages to present
this menu. This can also be faster, since double-tapping doesn't have
the same time delay that long press gestures have.