Turn libsignal-net request shadowing back down to "low" for beta users
Since the threshold for reporting is an absolute count of errors,
eventually everybody on the "high" shadowing mode will see them.
Rather than make that tracking more complex right now, turn the
shadowing back down to "low", which only ever logs about issues and
does not ask the user to report them to us. We've already gotten
the signal we need.