Fix incorrect section indices in chat list
If a new section happened to be inserted at the beginning of the chat list, all
the pinned & unpinned chat sections can have their section indices changed.
This was *almost* handled correctly, except when a chat is changing in the same
transaction as a section is being inserted or removed.

The fix is to dynamically compute the section index from either the new or old
CLVRenderState, depending on the table view change operation being animated.