Fix crash when opening GIF
I discovered this while re-testing after merging in the recent changes. If Reduce Motion is *off*, opening a GIF would crash because the app couldn’t set up playback controls on the GIF player. This fix is starting to give me a whiff of code smell though, so I might need to refactor some of this…