OWSChatConnection: De-promisify libsignal requests
There *is* a functionality change here: handling of responses and
request errors no longer jumps to the OWSChatConnection serial queue.
None of the processing done in either the success or failure case
cares what queue it's on, and because of the use of async there were
already no guarantees about the order of enqueues.

Groundwork for the next commit.