Improve Stale Issue and Stale PR processing by Stale Bot
Currently, all issues and PRs will get closed after 90 days of
inactivity. This means real issues get dismissed because nobody found
the time to take a closer look at it. This can be stopped by commenting
anything into the issue withing 7 days.

This leads duplicated issues and an issue tracker that is not
representative of the actual state and unhappy users. In some cases the
comment to keep the issue open cannot be done in time because 7 days is
not a lot of time and can easily by missed by a week of absence due to
holidays or sickness.

To fix this, this adjusts how the stale bot handles issues.
First off, issues/PRs labeled `acknowledged` issues will still be
ignored by stale bot and never closed.
Stale bot will only look at issues marked `answered` or `needs-info` and
PRS marked `needs-changes`. This means when a maintainer sees the issue
there are a few options:

- Mark the issue/PR as `acknowledged`, if it is an issue
- Answer the issue and mark the issue as `answered`, if the issue should
  be solved with the answer (e.g. support requests)
- Answer the issue or PR with request for more information or request
  changes and mark the issue as `needs-info` or the PR as
  `needs-changes`

This also improves the stale bot messages for marking stale with more
information about what is happening and what will happen.

The time from marking stale to closing the issue/PR is increasing to fit
two weeks, one weekend, and one extra day to allow people to live their
lives and go on holiday without the need of checking their emails.

This fixes #5931.