Fix unexpected highlighted area under "scroll down" button in chat on iOS 26.
I've previously made the buttons bigger (easier to tap) and added some padding around their "glass" backgrounds to preserve appearance spec (40 dp diamater). Turns out UIKit doesn't expect us to do that and dims the entire button with some shade of white on tap.

The fix was to make the button as small as it appears to the eye.